From structured to OO style.
The current style of the code is in a procedural style. "Here's a
function, it's namespaced by my class." The standard idiom would
have been to do this:

```
module Geolocator
  extend self

  def ip_lookup
  <snip>
end
```

However, Ruby is OO, not procedural. Let's take advantage of that!
Basically, you can construct an object with the arguments instead.
So now, we construct a geolocator object with an IP address, and
then ask it to look up the IP.

Right now it might not seem like much, but imagine two or three
operations on an object:

```
Geolocator.ip_lookup("123.456.789.123")
Geolocator.resolve_domain("123.456.789.123")
Geolocator.valid_ip?("123.456.789.123")
```

vs

```
locator = Geolocator.new("123.456.789.123")
locator.ip_lookup
locator.resolve_domain
locator.valid_ip?
```

Much more DRY.

There are also testing benefits to doing things this way: a bag
of methods module is harder to test than an object.