Getting rid of private by introducing new class.
Your original code felt the tension of two responsibilities: you
have your own processing to do, as well as processing from calling
an external service. You can see it in the code: you split it up
into two methods for a reason. However, you marked the second as
private, because you didn't want to make that part of your external
API.

That's great, but testing private methods is difficult, as your tests
show. You're testing the private method through the public one, which
is better than some Object#send shennanigans to test it directly.
However, we can do better!

Instead, let's make a new object, and consider it private, for our
own use. We can use that object to wrap the external API, which
provides two benefits:

1) clear respect of the SRP principle. One object for our stuff, one
object for external processing.
2) a seam to migrate to a different service in the future. We can implement
and use a different external service object, and none of our code
has to change.

Because we're testing through the public methods, no changes in tests are required.