using rspec's let feature to great effect.
Two changes here: the first is to use an awesome feature of rspec: let.

Basically, we can let rspec instantiate our object, which gets
all of that typing of long class names out of the way. It also lets
our tests focus on what's important, and ignore what isn't. Read the tests
now: "When the ip is this, I expect the locator to do that."

The second thing was breaking the last test into two. I wanted to do
this as a separate step, but it didn't work until I made the change:
you really had two assertions in that last one. It deserves its own
test.